--- a/client/src/App.css
+++ b/client/src/App.css
@@ -1,4 +1,4 @@
-/*! Todo Application - CSS Version 0.4.7 Copyright 2021, Stanley Owen */
+/*! Todo Application - CSS Version 0.4.8 Copyright 2021, Stanley Owen */
 @import url('https://fonts.googleapis.com/css2?family=Chilanka&display=swap');
 @import url('https://fonts.googleapis.com/css2?family=Noto+Sans+JP&display=swap');
 * {