--- a/client/src/components/account.component.js
+++ b/client/src/components/account.component.js
@@ -12,15 +12,20 @@ const SERVER_URL = process.env.REACT_APP_SERVER_URL;
 
 const Account = ({ userData }) => {
     const {email, id, thirdParty, verified, authenticated, isLoading} = userData;
-    const [oldPassword, setOldPassword] = useState();
-    const [newPassword, setNewPassword] = useState();
-    const [confirmPassword, setConfirmPassword] = useState();
     const [disabled, setDisabled] = useState(false);
-    const [visible, setVisibility] = useState({
+    const [data, setData] = useState({
+        oldPassword: '',
+        newPassword: '',
+        confirmPassword: ''
+    })
+    const [properties, setProperties] = useState({
         password: false,
         newPassword: false,
         confirmPassword: false
     })
+
+    const handleChange = (a, b) => setProperties({ ...properties, [a]: b })
+    const handleData = (a, b) => setData({ ...data, [a]: b })
 
     useEffect(() => {
         const background = document.getElementById('background');
@@ -54,21 +59,20 @@ const Account = ({ userData }) => {
         const btn = document.getElementById('change-password');
         async function submitData() {
             btn.innerHTML = "Updating..."; btn.setAttribute("disabled", "true"); btn.classList.add("disabled"); setDisabled(true);
-            await axios.put(`${SERVER_URL}/account/user`, { oldPassword, newPassword, confirmPassword }, { headers: { 'XSRF-TOKEN': getCSRFToken() }, withCredentials: true })
+            await axios.put(`${SERVER_URL}/account/user`, data, { headers: { 'XSRF-TOKEN': getCSRFToken() }, withCredentials: true })
             .then(res => {
                 closeModal('background', 'modal');
-                setOldPassword(''); setNewPassword(''); setConfirmPassword('');
+                setData({ oldPassword: '', newPassword: '', confirmPassword: '' })
                 setNotification(NOTIFICATION_TYPES.SUCCESS, res.data.message);
             })
             .catch(err => setNotification(NOTIFICATION_TYPES.DANGER, err.response.data.message));
             btn.innerHTML = "Update"; btn.removeAttribute("disabled"); btn.classList.remove("disabled"); setDisabled(false);
         }
-        if(!oldPassword || !newPassword || !confirmPassword){ setNotification(NOTIFICATION_TYPES.DANGER, "Please Make Sure to Fill Out All Required the Fields !"); document.getElementById(!oldPassword ? 'old-password' : !newPassword ? 'new-password' : 'confirm-password').focus(); }
-        else if(oldPassword.length < 6 || newPassword.length < 6 || confirmPassword.length < 6 || oldPassword.length > 40 || newPassword.length > 40 || confirmPassword.length > 40){ setNotification(NOTIFICATION_TYPES.DANGER, 'Please Provide a Password between 6 ~ 40 characters !'); document.getElementById(oldPassword.length < 6 || oldPassword.length > 40 ? 'old-password' : newPassword.length < 6 || newPassword.length > 40 ? 'new-password' : 'confirm-password').focus(); }
-        else if(newPassword !== confirmPassword) { setNotification(NOTIFICATION_TYPES.DANGER, 'Please Make Sure Both Passwords are Match !'); document.getElementById('confirm-password').focus(); }
+        if(!data.oldPassword || !data.newPassword || !data.confirmPassword){ setNotification(NOTIFICATION_TYPES.DANGER, "Please Make Sure to Fill Out All Required the Fields !"); document.getElementById(!data.oldPassword ? 'old-password' : !data.newPassword ? 'new-password' : 'confirm-password').focus(); }
+        else if(data.oldPassword.length < 6 || data.newPassword.length < 6 || data.confirmPassword.length < 6 || data.oldPassword.length > 40 || data.newPassword.length > 40 || data.confirmPassword.length > 40){ setNotification(NOTIFICATION_TYPES.DANGER, 'Please Provide a Password between 6 ~ 40 characters !'); document.getElementById(data.oldPassword.length < 6 || data.oldPassword.length > 40 ? 'old-password' : data.newPassword.length < 6 || data.newPassword.length > 40 ? 'new-password' : 'confirm-password').focus(); }
+        else if(data.newPassword !== data.confirmPassword) { setNotification(NOTIFICATION_TYPES.DANGER, 'Please Make Sure Both Passwords are Match !'); document.getElementById('confirm-password').focus(); }
         else submitData();
     }
-    const handleChange = (a, b) => setVisibility({ ...visible, [a]: !b });
 
     const closeInfo = (a) => {
         const infos = document.getElementById('infos')
@@ -155,10 +159,10 @@ const Account = ({ userData }) => {
                                 <div className="m-10">
                                     <div className="contact__infoField">
                                         <label htmlFor="old-password">Old Password <span className="required">*</span></label>
-                                        <input title="Old Password" id="old-password" type={ visible.password ? 'text':'password' } className="contact__inputField" onChange={(event) => setOldPassword(event.target.value)} value={oldPassword} spellCheck="false" autoCapitalize="none" required autoComplete={ visible.password ? 'off':'current-password'} />
+                                        <input title="Old Password" id="old-password" type={ properties.password ? 'text':'password' } className="contact__inputField" onChange={(event) => handleData('oldPassword', event.target.value)} value={data.oldPassword} spellCheck="false" autoCapitalize="none" required autoComplete={ properties.password ? 'off':'current-password'} />
                                         <span className="contact__onFocus"></span>
-                                        <IconButton className="view-eye" onClick={() => handleChange('password', visible.password)}>
-                                            <FontAwesomeIcon icon={visible.password ? faEyeSlash : faEye} />
+                                        <IconButton className="view-eye" onClick={() => handleChange('password', !properties.password)}>
+                                            <FontAwesomeIcon icon={properties.password ? faEyeSlash : faEye} />
                                         </IconButton>
                                     </div>
                                 </div>
@@ -166,20 +170,20 @@ const Account = ({ userData }) => {
                                     <div className="m-10">
                                         <div className="contact__infoField">
                                             <label htmlFor="new-password">New Password <span className="required">*</span></label>
-                                            <input title="New Password" id="new-password" type={ visible.newPassword ? 'text':'password' } className="contact__inputField" onChange={(event) => setNewPassword(event.target.value)} value={newPassword} spellCheck="false" autoCapitalize="none" required autoComplete={ visible.newPassword ? 'off':'new-password'} />
+                                            <input title="New Password" id="new-password" type={ properties.newPassword ? 'text':'password' } className="contact__inputField" onChange={(event) => handleData('newPassword', event.target.value)} value={data.newPassword} spellCheck="false" autoCapitalize="none" required autoComplete={ properties.newPassword ? 'off':'new-password'} />
                                             <span className="contact__onFocus"></span>
-                                            <IconButton className="view-eye" onClick={() => handleChange('newPassword', visible.newPassword)} name="newPassword">
-                                                <FontAwesomeIcon icon={visible.newPassword ? faEyeSlash : faEye} />
+                                            <IconButton className="view-eye" onClick={() => handleChange('newPassword', !properties.newPassword)} name="newPassword">
+                                                <FontAwesomeIcon icon={properties.newPassword ? faEyeSlash : faEye} />
                                             </IconButton>
                                         </div>
                                     </div>
                                     <div className="m-10">
                                         <div className="contact__infoField">
                                             <label htmlFor="confirm-password">Confirm New Password <span className="required">*</span></label>
-                                            <input title="Confirm New Password" id="confirm-password" type={ visible.confirmPassword ? 'text':'password' } className="contact__inputField" onChange={(event) => setConfirmPassword(event.target.value)} value={confirmPassword} spellCheck="false" autoCapitalize="none" required autoComplete={ visible.confirmPassword ? 'off':'new-password'} />
+                                            <input title="Confirm New Password" id="confirm-password" type={ properties.confirmPassword ? 'text':'password' } className="contact__inputField" onChange={(event) => handleData('confirmPassword', event.target.value)} value={data.confirmPassword} spellCheck="false" autoCapitalize="none" required autoComplete={ properties.confirmPassword ? 'off':'new-password'} />
                                             <span className="contact__onFocus"></span>
-                                            <IconButton className="view-eye" onClick={() => handleChange('confirmPassword', visible.confirmPassword)} name="confirmPassword">
-                                                <FontAwesomeIcon icon={visible.confirmPassword ? faEyeSlash : faEye} />
+                                            <IconButton className="view-eye" onClick={() => handleChange('confirmPassword', !properties.confirmPassword)} name="confirmPassword">
+                                                <FontAwesomeIcon icon={properties.confirmPassword ? faEyeSlash : faEye} />
                                             </IconButton>
                                         </div>
                                     </div>