… README.md
… client/package.json
… client/src/App.min.css
… client/src/components/login.component.js
… client/src/components/logout.component.js
… client/src/components/navbar.component.js
… client/src/components/register-oauth.component.js
… client/src/components/register.component.js
… client/src/components/req-oauth.component.js
… client/src/components/reset-password.component.js
… client/src/components/verify-account.component.js
… client/src/img/verified-error.gif
… client/src/img/verified-success.gif
… client/src/img/verified-success.mp4
… client/src/libraries/validation.js
… package.json
… server/index.js
… server/lib/callback.js
… server/lib/crypto.js
… server/lib/passport.js
… server/models/otp-token.model.js
… server/models/revoke-token.model.js
… server/models/users.model.js
… server/routes/oauth.route.js
… server/routes/users.route.js
… server/template.html