**What's Changed in v0.4.8:**
1. Added Multi Factor Authentication (0cbd8373aa7daf89ad3fc928a62651aac8821456) (`MFA`) (`Beta`):
    - Added 2FA Model (59dce41dac6e062e918a307afb357736190bed72)
    - Added `Send OTP` Feature in `Server` (9cac69e2709bda88819920a48eb6f2db4ecf4bd5)
    - Added Verify `OTP` Feature (ada228e3e7a4691af41c7b77297b9215596a59bc) (3a7a02f3c6621978ad8bb15502c4a209ae611fb2)
    - Added Resend OTP in `Login` Component (adb231c9975d89612d7a82769ba3f86471444fde)
    - Added `Logout` Component (523e55c9334192aed6d75bcb8ec4f9187a3ab718)
    - Added Re-Fetch Data if Error Occurred Feature (#113) (64fc7ec4d80e6f897d252ca5bdb96a38604edbf2)
    - Added Enable 2FA on `Client` Side (7ecf5b6d282de7c6164f5ba661f21ab150e3b224)
    - Added `Activate` and `Deactivate` `MFA` on `Server` Side (64855e1b00077728c55cd3a5b8d2f9027f2e8507)
2. Optimized Code Quality (59cdd4d436acf70cb915a8a04e7b75402c694a9a)
3. Unverified `2FA` Users will be Redirected to `Login` Component (1038070f5fafe514fa1df2f7d160ddf7a4f6e027)
4. Email will now Automatically Appear on Forgot Password if user is Logged In (c748378fe3ef227606c1a6a17a985f8020259c35)
5. Isolate Each Secret Key for OTP and Token (#110) (86d7c0f2e4ec4855805c37fe5ab5c4297b7460b7)
6. Improved Performance (dd3210f30297e3b79e46721baa6b0bf81f7988d5)
7. Improved `UI` and `UX` (234e315559982b4a9dc03c18d46dc425231701e8)
8. Upgrade `Dependencies` to the Latest Version (e406ecc378fabd99bbff87240bb28e6e082c3abd)
9. Resolved Issues of:
    - Inconsistence Dark Mode (#111) (2b2709a366b0aae376826bb0c5356aa33b894874)
    - Being Redirected After `Resend OTP` is Clicked (#112) (9802c61a73ea267e9dc168246a573c3833390a59)
    - Error in `OAuth` and `Home` Components (baac69147c9a5caf0919900dc02743b5d1a19ed4)
    - Minor Issues

Details About Changes in v0.4.8:
| Client | `build\static\js\2.6ca8d93c.chunk.js` | `build\static\js\main.2591d28a.chunk.js` | `build\static\css\main.030489ff.chunk.css` | `build\static\js\3.668acfa5.chunk.js` | `build\static\js\runtime-main.3cbd5898.js`
------ | --------------------------------------- | ------------------------------------------- | ---------------------------------------------- | -------------------------------------------------- | ---------------------------------------------- |
| v0.4.8 | 148.95 kB (+42.58 kB) | 27.81 kB (+2.2 kB) | 4.5 kB (+224 B) | 1.56 kB (+2 B) | 1.16 kB (+2 B) |
| v0.4.7 | 148.29 kB (-1.25 kB) | 26.54 kB (+478 B) | 4.47 kB (+72 B) | 1.56 kB | 1.16 kB |
| v0.4.6 | 147.82 kB | 26.07 kB | 4.4 kB | 1.56 kB | 1.16 kB |
| v0.4.5 | 106.38 kB | 25.61 kB | 4.29 kB | 1.56 kB | 1.16 kB |

| Server | Size |
| ------- | ---- |
| v0.4.8 | 211 kB (+21 kB) |
| v0.4.7 | 190 kB (+14 kB) |
| v0.4.6 | 176 kB |

For more information about changes in this release, please visit our [project roadmaps v0.4.8](https://github.com/stanleyowen/todo-application/projects/28)